#29 Move tag to cost
