+++ lib/domain/cost/cost.dart
@@ -2,7 +2,7 @@ import 'package:flutter_template/domain/cost/amount.dart';
 import 'package:flutter_template/domain/cost/category.dart';
 import 'package:flutter_template/domain/cost/point.dart';
 import 'package:flutter_template/domain/cost/title.dart';
-import 'package:flutter_template/domain/tag/tags.dart';
+import 'package:flutter_template/domain/cost/tags.dart';
 import 'package:freezed_annotation/freezed_annotation.dart';
 import 'package:uuid/uuid.dart';
