+++ lib/presentation/features/registration/registration_page.dart
@@ -3,7 +3,7 @@ import 'package:flutter/services.dart';
 import 'package:flutter_hooks/flutter_hooks.dart';
 import 'package:flutter_template/domain/cost/category.dart';
 import 'package:flutter_template/domain/cost/point.dart';
-import 'package:flutter_template/domain/tag/tag.dart';
+import 'package:flutter_template/domain/cost/tag.dart';
 import 'package:flutter_template/presentation/features/home/home_page.dart';
 import 'package:flutter_template/presentation/features/registration/registration_view_model.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
@@ -19,7 +19,7 @@ class RegistrationPage extends HookConsumerWidget {
     var tagTextController = useTextEditingController();
 
     var handleEnterTag = useCallback((value) {
-      viewModel.addTag(Tag.initial(value));
+      viewModel.addTag(Tag.of(value));
       tagTextController.clear();
     }, [viewModel, tagTextController]);
     var handleDeleteTag = useCallback((Tag tag) {
