+++ lib/presentation/features/registration/registration_state.dart
@@ -1,6 +1,6 @@
 import 'package:flutter_template/domain/cost/category.dart';
 import 'package:flutter_template/domain/cost/point.dart';
-import 'package:flutter_template/domain/tag/tags.dart';
+import 'package:flutter_template/domain/cost/tags.dart';
 import 'package:freezed_annotation/freezed_annotation.dart';
 
 part 'registration_state.freezed.dart';
