+++ lib/presentation/features/registration/registration_view_model.dart
@@ -4,8 +4,8 @@ import 'package:flutter_template/domain/cost/cost.dart';
 import 'package:flutter_template/domain/cost/cost_repository.dart';
 import 'package:flutter_template/domain/cost/point.dart';
 import 'package:flutter_template/domain/cost/title.dart';
-import 'package:flutter_template/domain/tag/tag.dart';
-import 'package:flutter_template/domain/tag/tags.dart';
+import 'package:flutter_template/domain/cost/tag.dart';
+import 'package:flutter_template/domain/cost/tags.dart';
 import 'package:flutter_template/presentation/features/registration/registration_state.dart';
 import 'package:flutter_template/infrastructure/local_storage/domain/cost/cost_db_repository.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
