+++ test/domain/cost/model_support.dart
@@ -2,8 +2,8 @@ import 'package:flutter_template/domain/cost/amount.dart';
 import 'package:flutter_template/domain/cost/category.dart';
 import 'package:flutter_template/domain/cost/cost.dart';
 import 'package:flutter_template/domain/cost/point.dart';
-import 'package:flutter_template/domain/tag/tag.dart';
-import 'package:flutter_template/domain/tag/tags.dart';
+import 'package:flutter_template/domain/cost/tag.dart';
+import 'package:flutter_template/domain/cost/tags.dart';
 import 'package:flutter_template/domain/cost/title.dart';
 
 class CostBuilder {
